+++ investpor/core/SourceFiles/cryptocurrencytransaction.cpp
@@ -4,14 +4,14 @@ namespace investpor {
 
     namespace core {
 
-        CryptocurrencyTransaction::CryptocurrencyTransaction(int id) :
+        CryptocurrencyTransaction::CryptocurrencyTransaction(quint16 id) :
             Transaction(id)
         {
 
         }
 
         CryptocurrencyTransaction::CryptocurrencyTransaction(Operation op, Cryptocurrency ccurrency,
-            double prc, double amnt, QDateTime opDateTime, double gPrice, int id) :
+            double prc, double amnt, QDateTime opDateTime, double gPrice, quint16 id) :
             Transaction(id), operationType(op), cryptoCurrency(ccurrency), price(prc),
             amount(amnt), operationDateTime(opDateTime), goalPrice(gPrice)
         {
